mv get_all_download_formats to resource.py
